% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{calc_stat_normal}
\alias{calc_stat_normal}
\title{Calculate summary metrics for a numeric vector}
\usage{
calc_stat_normal(x, ci = 0.95, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A dataframe containing summary metrics
}
\description{
This function assumes the numbers follow normal distribution
}
\examples{
add(1, 1)
add(10, 1)
}
